+++ CHANGELOG.md
@@ -16,4 +16,6 @@
 * [UIBULKED-11](https://issues.folio.org/browse/UIBULKED-11) Bulk Edit app - permissions 
 * [UIBULKED-19](https://issues.folio.org/browse/UIBULKED-19) Download errors (CSV)
 * [UIBULKED-17](https://issues.folio.org/browse/UIBULKED-17) Start bulk edit
+* [UIBULKED-18](https://issues.folio.org/browse/UIBULKED-18) Download matched records (CSV)
+* [UIBULKED-15](https://issues.folio.org/browse/UIBULKED-15) Add Query textarea
 * [UIBULKED-12](https://issues.folio.org/browse/UIBULKED-12) Populating preview of matched records based on identifiers search